+++ frontend/src/components/MoSearchBar/MoSearchBar.vue
@@ -2,21 +2,11 @@ SPDX-FileCopyrightText: 2017-2020 Magenta ApS
 SPDX-License-Identifier: MPL-2.0
 <template>
   <div class="search">
-    <div class="input-group">
-      <div class="input-group-prepend">
-        <span class="input-group-text"><icon name="search"/></span>
-      </div>
-      <v-autocomplete
-        :items="orderedListOptions"
-        v-model="item"
-        :get-label="getLabel"
-        :component-item="template"
-        @item-selected="selected"
-        @update-items="updateItems"
-        :auto-select-one-item="false"
-        :min-len="2"
-        :placeholder="$t('common.search')"
-        class="search-bar"
+    <div class="input-group input">
+      <mo-autocomplete
+        :search="updateItems"
+        :getResultValue="getResultValue"
+        :onSubmit="selected"
       />
     </div>
     <div class="input-group date-input">
@@ -36,9 +26,7 @@ SPDX-License-Identifier: MPL-2.0
 
 import sortBy from 'lodash.sortby'
 import Search from '@/api/Search'
-import VAutocomplete from 'v-autocomplete'
-import 'v-autocomplete/dist/v-autocomplete.css'
-import MoSearchBarTemplate from './MoSearchBarTemplate'
+import MoAutocomplete from '@/components/MoAutocomplete/MoAutocomplete.vue'
 import { MoInputDate } from '@/components/MoInput'
 import { AtDate } from '@/store/actions/atDate'
 
@@ -46,43 +34,19 @@ export default {
   name: 'MoSearchBar',
 
   components: {
-    VAutocomplete,
-    MoInputDate
+    MoAutocomplete,
+    MoInputDate,
   },
 
   props: {
-    'hideDateInput': Boolean
+    hideDateInput: Boolean,
   },
 
   data () {
     return {
-      /**
-       * The item, items, routeName component value.
-       * Used to detect changes and restore the value.
-       */
       item: null,
-      items: [],
       routeName: '',
-
       atDate: new Date(),
-
-      /**
-       * The template component value.
-       * Used to add MoSearchBarTemplate to the v-autocomplete.
-       */
-      template: MoSearchBarTemplate,
-
-      /**
-       * The noItem component value.
-       * Used to give a default name.
-       */
-      noItem: [{ name: this.$t('alerts.no_search_results') }]
-    }
-  },
-
-  computed: {
-    orderedListOptions () {
-      return sortBy(this.items, 'name')
     }
   },
 
@@ -118,13 +82,6 @@ export default {
 
   methods: {
     /**
-     * Get label name.
-     */
-    getLabel (item) {
-      return item ? item.name : null
-    },
-
-    /**
      * Get to the route name.
      * So if we're viewing an employee, it goes to the employee detail.
      */
@@ -142,20 +99,30 @@ export default {
      */
     updateItems (query) {
       let vm = this
-      vm.items = []
       let org = this.$store.state.organisation
-      if (vm.routeName === 'EmployeeDetail') {
-        Search.employees(org.uuid, query)
-          .then(response => {
-            vm.items = response.length > 0 ? response : vm.noItem
-          })
-      }
-      if (vm.routeName === 'OrganisationDetail') {
-        Search.organisations(org.uuid, query, this.atDate)
-          .then(response => {
-            vm.items = response.length > 0 ? response : vm.noItem
-          })
-      }
+
+      return new Promise(resolve => {
+        if (query.length < 2) {
+          return resolve([])
+        }
+
+        if (vm.routeName === 'EmployeeDetail') {
+          Search.employees(org.uuid, query)
+            .then(response => {
+              resolve(sortBy(response, 'name'))
+            })
+        }
+        if (vm.routeName === 'OrganisationDetail') {
+          Search.organisations(org.uuid, query, this.atDate)
+            .then(response => {
+              resolve(sortBy(response, 'name'))
+            })
+        }
+      })
+    },
+
+    getResultValue (result) {
+      return result.name
     },
 
     /**
@@ -183,15 +150,6 @@ export default {
     align-items: center; /* vertically center items inside input group */
     flex-wrap: nowrap;
     width: auto;
-  }
-  .search .input-group input {
-    width: 1%;
-  }
-  .search .input-group .input-group-prepend {
-    flex: unset;
-  }
-  .search .input-group .input-group-prepend .input-group-text {
-    display: inline;
   }
   .search .input-group.date-input {
     max-width: 10vw;